# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: One Bharat Export-Import API
  description: >
    REST API for the One Bharat Export-Import platform — importer/exporter/logistics
    trade workflow with escrow payments, shipment tracking, RFQ/quotation, chat,
    trade document generation, and platform revenue features.
    Real-time chat additionally uses a WebSocket endpoint (GET /api/v1/ws/chat?token=...)
    not representable as a standard HTTP operation below.
  version: "1.0.0"
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /fleet:
    post:
      tags: [Fleet]
      summary: Add a vehicle to my fleet (logistics only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vehicle_type, registration_number]
              properties:
                vehicle_type: { type: string }
                registration_number: { type: string }
                capacity_kg: { type: number, nullable: true }
                capacity_cbm: { type: number, nullable: true }
                service_route: { type: string, nullable: true }
                status: { type: string, enum: [active, inactive, maintenance] }
      responses:
        '201': { description: Vehicle added }
...
  /fleet/{id}:
    put:
      tags: [Fleet]
      summary: Update a fleet vehicle (logistics only, must own it)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vehicle_type: { type: string }
                registration_number: { type: string }
                capacity_kg: { type: number, nullable: true }
                capacity_cbm: { type: number, nullable: true }
                service_route: { type: string, nullable: true }
                status: { type: string }
      responses:
        '200': { description: Vehicle updated }
    delete:
      tags: [Fleet]
      summary: Remove a fleet vehicle (logistics only, must own it)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Vehicle removed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT